refactor: move spatial operator into instruments package
